pyproject.toml
[tool.poetry]
name = "is-project-thailand-tourism-recommendation"
version = "0.1.0"
description = ""
authors = ["Ittichai Boonyarakthunya <i.boonyarakthunya@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.12,<3.13"
pandas = ">=1.3.5,<2.2.1"
scikit-learn = "^1.5.1"
requests = "^2.32.3"
python-dotenv = "^1.0.1"
matplotlib = "^3.9.1"
geopy = "^2.4.1"
folium = "^0.17.0"
geopandas = "^1.0.1"
urllib3 = "^2.2.2"
certifi = "^2024.7.4"
taipy = "^3.1.1"
jupyter = "^1.0.0"
pythainlp = "^5.0.4"
seaborn = "^0.13.2"
imblearn = "^0.0"
parsel = "^1.9.1"
bs4 = "^0.0.2"
nltk = "^3.9.1"
spicy = "^0.16.0"
torch = "^2.4.0"
torchvision = "^0.19.0"
tensorboard = ">=2.17.0"
tensorboard-data-server = "0.7.2"
tensorflow = "2.17.0"
tensorflow-datasets = "4.9.6"
tensorflow-hub = "0.16.1"
tensorflow-io-gcs-filesystem = "0.37.1"
tensorflow-metadata = "1.15.0"
tensorflow-probability = "0.24.0"
tensorstore = "0.1.64"
scikeras = "^0.13.0"
gensim = "^4.3.3"
pydot = "^3.0.1"
graphviz = "^0.20.3"
shap = "^0.46.0"
emoji = "^2.12.1"
tweepy = "^4.14.0"
surprise = "^0.1"

[tool.poetry.group.dev.dependencies]
ipykernel = "^6.29.5"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
